Relocate factor_interaction.py to formula/transforms/
Pure rename — no logic changes. Moves factor_interaction.py into the
new transforms/ subpackage and updates the import in model_matrix.py.
Prepares the directory layout for additional transform modules.
