+++ pyfixest/estimation/formula/transforms/factor_interaction.py
